Extract `useTimeInput` hook
